+++ internal/impl/gcp/cache_cloud_storage.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"cloud.google.com/go/storage"
+	"google.golang.org/api/iterator"
 
 	"github.com/warpstreamlabs/bento/public/service"
 )
@@ -127,6 +128,22 @@ func (c *gcpCloudStorageCache) Add(ctx context.Context, key string, value []byte
 
 func (c *gcpCloudStorageCache) Delete(ctx context.Context, key string) error {
 	return c.bucketHandle.Object(key).Delete(ctx)
+}
+
+func (c *gcpCloudStorageCache) ListKeys(ctx context.Context) ([]string, error) {
+	var keys []string
+	it := c.bucketHandle.Objects(ctx, nil)
+	for {
+		attrs, err := it.Next()
+		if errors.Is(err, iterator.Done) {
+			break
+		}
+		if err != nil {
+			return nil, err
+		}
+		keys = append(keys, attrs.Name)
+	}
+	return keys, nil
 }
 
 func (c *gcpCloudStorageCache) Close(ctx context.Context) error {